Refactor: Consolidate duplicated utility functions
This commit addresses code duplication in utility and API endpoint files:

- Consolidated `api_response`:
    - Removed `api_response` from `utils/authutils.py`.
    - Updated imports to use `api_response` from `utils/responseutils.py`.

- Consolidated session validation logic:
    - Removed `is_session_valid`, `invalidate_session`, `validate_session`,
      and related global variables from `utils/authutils.py`.
    - Updated imports to use these functions from `utils/sessionutils.py`.

- Centralized validation functions:
    - Moved `validate_username` and `validate_password` from
      `apiendpoints/auth.py` to `utils/authutils.py`.
    - Updated imports in `apiendpoints/auth.py`, `apiendpoints/user.py`,
      and `apiendpoints/password.py` accordingly.

These changes improve code maintainability and reduce redundancy. I performed a manual code review as I couldn't find any automated tests.
