+++ utils/authutils.py
@@ -1,28 +1,41 @@
 from utils.userutils import get_user
 from functools import wraps
 from flask_jwt_extended import get_jwt_identity, verify_jwt_in_request
-from utils.logutils import log_error
+import re # Added import for validate_username
+from utils.logutils import log_error, log_info
 from flask import jsonify
+from .responseutils import api_response
+from .sessionutils import is_session_valid # Added import
 
-def api_response(success, message=None, data=None, status_code=200):
-    """
-    Standardized API response format
+# Moved from apiendpoints/auth.py
+def validate_username(username):
+    """Validate username against security requirements"""
+    if not username or not isinstance(username, str):
+        return False, "Username must be a non-empty string"
     
-    Args:
-        success: Boolean indicating success/failure
-        message: Optional message string
-        data: Optional data dictionary
-        status_code: HTTP status code
-        
-    Returns:
-        tuple: (Flask response object, status_code)
-    """
-    response = {'success': success}
-    if message:
-        response['message'] = message
-    if data:
-        response.update(data)
-    return jsonify(response), status_code
+    if len(username) < 3:
+        return False, "Username must be at least 3 characters long"
+    
+    if not re.match(r'^[a-zA-Z0-9_]+$', username):
+        return False, "Username can only contain letters, numbers, and underscores"
+    
+    return True, None
+
+def validate_password(password):
+    """Validate password against security requirements"""
+    if not password or not isinstance(password, str):
+        return False, "Password must be a non-empty string"
+    
+    if len(password) < 6:
+        return False, "Password must be at least 6 characters long"
+    
+    if not any(char.isdigit() for char in password):
+        return False, "Password must contain at least one number"
+    
+    if not any(char.isalpha() for char in password):
+        return False, "Password must contain at least one letter"
+    
+    return True, None
 
 def auth_middleware():
     """
@@ -70,65 +83,4 @@ def admin_required(f):
         return f(*args, **kwargs)
     return decorated
 
-# Session validation functions
-invalid_sessions = set()
-exempt_sessions = {}
-
-def is_session_valid(username):
-    """
-    Check if a user session is valid
-    
-    Args:
-        username: Username to check
-        
-    Returns:
-        Boolean indicating if session is valid
-    """
-    from flask import request
-    
-    user = get_user(username)
-    
-    if not user:
-        return False
-    
-    current_sid = None
-    if hasattr(request, 'sid'):
-        current_sid = request.sid
-    
-    user_id_str = str(user['id'])
-    if user_id_str in invalid_sessions:
-        if user_id_str in exempt_sessions and exempt_sessions[user_id_str] == current_sid:
-            return True
-        return False
-    
-    if user.get('is_suspended', False):
-        return False
-    
-    return True
-
-def invalidate_user_session(user_id):
-    """
-    Invalidate a user session by ID
-    
-    Args:
-        user_id: User ID to invalidate
-    """
-    user_id_str = str(user_id)
-    invalid_sessions.add(user_id_str)
-    if user_id_str in exempt_sessions:
-        del exempt_sessions[user_id_str]
-    log_info(f"Session invalidated for user ID: {user_id}")
-
-def validate_user_session(user_id):
-    """
-    Validate a previously invalidated session
-    
-    Args:
-        user_id: User ID to validate
-    """
-    user_id_str = str(user_id)
-    if user_id_str in invalid_sessions:
-        invalid_sessions.remove(user_id_str)
-    if user_id_str in exempt_sessions:
-        del exempt_sessions[user_id_str]
-    log_info(f"Session validated for user ID: {user_id}")
+# Session validation functions removed from here
